refactor: `UpdateReminder`
