+++ Backend/src/Modules/Reminders/Reminders.Api/Controllers/RemindersRequests.cs
@@ -12,3 +12,12 @@ public record CreateReminderRequest
     public string OwnerId { get; set; }
     public NodeId NodeId { get; set; }
 }
+
+public record UpdateReminderRequest
+{
+    public string Title { get; set; }
+    public string Description { get; set; }
+    public DateTime NotifyAt { get; set; }
+    public int Priority { get; set; }
+    public NodeId NodeId { get; set; }
+}
